Improve backport logic - support explicitly-referenced non-maintenance branches
When we backport with an "explicit" branch reference, e.g. `backport to 5.4`, internally the action is iterating through the list of [`supported-maintenance-versions`](https://github.com/hazelcast/hazelcast/blob/master/.github/actions/get-supported-maintenance-versions/action.yml) to figure out if it's valid.

This _works_ but the `5.6` release exposed a workflow issue - until it's added to `supported-maintenance-versions` (_after_ it's released) - the backport label does nothing - because it wasn't on that list.

[In `hazelcast-mono`, the backport process works differently](https://github.com/hazelcast/hazelcast-mono/blob/master/.github/workflows/backport.yml) - we simply use the `backport to` prefix to extract the target branch, and then use it. It has no `supported-maintenance-versions` dependency.

This is more complicated to implement in docs because of the `backport to all versions` requirement, but it's still possible:
- `backport to all versions` case, we supply the exact label to match against, and the set of branches it should backport to
   - if the label is on the PR, then backport to the branches
- otherwise, we simply supply the prefix (e.g. `backport to`) and let the action derive the rest
   - for each label on the PR, the branch referenced (if present) - e.g. `backport to {branch}`
   - cross-check the branch actually exists to avoid getting confused with input `backport to all branches` (because this _could_ otherwise be interpreted as backporting to the `all branches` branch 😶)
   - run the backport for _each_ of the identified branches
- stop pruning the `backport` labels for non-supported versions - as any (arbritrary) version is now supported

This _also_ has the benefit of removing one of the blockers for merging the two differing implementations (docs / `hazelcast-mono`) in future.

Post-merge actions:
- [ ] update the backport label names to _exactly match_ the branch name (e.g. `backport to 5.4` -> `backport to v/5.4`)
- [ ] copy this fix to [every doc repo](https://github.com/search?q=org%3Ahazelcast%20%20%22backport-workflow%22&type=code)